\name{change.parm}
\alias{change.parm}
\title{
Change parameter scope.
}
\description{
Function to change the parameter scope.
}
\usage{
change.parm(object, listall = TRUE, classic = FALSE)
}
\arguments{
  \item{object}{
The xpose data opject.
}
  \item{listall}{
whether we should list all the current parameters.
}
  \item{classic}{
true if used in the classic menu system (for internal use).
}
}

\value{
 If classic then return nothing.  Otherwise return the new data object.
}
\author{
Andrew C. Hooker
}
\examples{
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{methods}

