\name{read.TTE.sim.data}
\alias{read.TTE.sim.data}

\title{
Read (repeated) time-to-event simulation data files.
}
\description{
Read (repeated) time-to-event simulation data files.
}
\usage{
read.TTE.sim.data(sim.file,
                  subset = NULL,
                  headers = c("REP", "ID", "DV", "TIME", "FLAG2","DOSE"),
                  xpose.table.file = FALSE,
                  ...)
}
\arguments{
  \item{sim.file}{
Name of the simulated file.
}
  \item{subset}{
subset to extract.
}
  \item{headers}{
headers in file.
}
  \item{xpose.table.file}{
xpose table files.
}
  \item{\dots}{
Extra arguments passed to function.
}
}
\details{
}
\value{
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
}
\references{
}
\author{
  Andrew C. Hooker
}
\note{
}

\seealso{
}
\examples{
}
\keyword{ methods }
